Finishing Auth + Creating Getting Started Page

+++ frontend/src/components/ExerciseCard.jsx
@@ -1,11 +1,20 @@
 import useExercisesContext from "../hooks/useExercisesContext"
+import useAuthContext from "../hooks/useAuthContext"
 
 export default function ExerciseCard ({ exercise }) {
     const { dispatch } = useExercisesContext()
+    const { user } = useAuthContext()
     
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+        
         const response = await fetch('http://localhost:4000/api/exercises/' + exercise._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
